[J:MEP-0] phase-3 step 3: rename auth sibling to nonWear; split projects driving into parent+nonWear+wear
Two coordinated refactors per plan D2/D5 + user direction on naming scope:

Auth rename (mobile -> nonWear):
- feat/authentication/fe/driven/mobile renamed to .../nonWear (sibling
  holds the phone-Android redirect-URI binding — naming aligns with the
  projects driving split where "nonWear" spans phone-Android + iOS + web
  + JVM). Module registration, androidApp dep, and MainActivity import
  updated. Package renamed from ...mobile.di to ...nonwear.di (lowercase
  per compose-resources/Kotlin package convention).

Projects driving split:
- feat/projects/fe/driving/impl retains ViewModels (ProjectsViewModel,
  ProjectDetailsViewModel/abstract + NonWeb/Web concrete variants,
  ProjectDetailsEditViewModel, ProjectAssignmentsViewModel), UiState,
  VmState, and Mappers. Tests stay in parent.
- New feat/projects/fe/driving/nonWear holds all Compose-Multiplatform
  Material 3 content: Composables (projects list + details + details
  edit + assignments screens, all ui/ subtrees), the Koin module with
  navigation registrations (including webMain + nonWebMain actuals for
  platformModule expect), the web/nonWeb platform SaveReportEffect
  actuals, and strings.xml. All moves via git rename.
- New feat/projects/fe/driving/wear is a scaffold; WearProjectList +
  Wear Koin module land in Phase 3 step 6.

Predicted findings realised (logged for critique md):
- ProjectsViewModel, ProjectDetailsViewModel, ProjectDetailsEditViewModel,
  ProjectAssignmentsViewModel, NonWebProjectDetailsViewModel,
  WebProjectDetailsViewModel widened from internal to public so the
  nonWear sibling Koin module can reference them (P2 falsification:
  non-rename parent-source edit).
- ProjectsUiState, ProjectDetailsUiState (+ ProjectDetailsUiStatus,
  StructuresUiStatus, InspectionsUiStatus enums), ProjectDetailsEditUiState,
  ProjectAssignmentsUiState, AssignedUserItem widened from internal to
  public for the same reason.
- ProjectDetailsViewModel.vmState and .errorEventsChannel downgraded
  from protected to internal to avoid leaking the still-internal
  ProjectDetailsVmState via a public parent class's protected property.
- Compose resources namespace in moved files updated from
  snag.feat.projects.fe.driving.impl.generated.resources to
  snag.feat.projects.fe.driving.nonwear.generated.resources — forced by
  the compose resources generator, which namespaces per-module (lowercased
  nonwear segment). Consumer code unchanged; imports internal to nonWear.
- Parent build.gradle.kts: driving:api deps for structures/clients/
  inspections/users dropped (only Koin module navigation used them; now
  in nonWear). Kept app:api deps as api scope so nonWear picks up model
  types transitively.
- koinModulesAggregate/fe/build.gradle.kts adds the nonWear sibling
  dependency so the aggregate's existing import of projectsDrivingImplModule
  resolves to the nonWear-hosted binding (same package name; different
  producing module).

Shared port/app/model layers across both features: untouched.
