+++ androidApp/build.gradle.kts
@@ -89,7 +89,7 @@ android {
 
 dependencies {
     implementation(projects.composeApp)
-    implementation(projects.feat.authentication.fe.driven.mobile)
+    implementation(projects.feat.authentication.fe.driven.nonWear)
     implementation(libs.koin.core)
     implementation(libs.oidc.appsupport)
     implementation(libs.androidx.activity.compose)
