--- feat/authentication/fe/driven/mobile/src/androidMain/kotlin/cz/adamec/timotej/snag/authentication/fe/driven/mobile/di/AuthenticationDrivenMobileModule.kt
@@ -1,22 +1,0 @@
-/*
- * Copyright (c) 2026 Timotej Adamec
- * SPDX-License-Identifier: MIT
- *
- * This file is part of the thesis:
- * "Multiplatform snagging system with code sharing maximisation"
- *
- * Czech Technical University in Prague
- * Faculty of Information Technology
- * Department of Software Engineering
- */
-
-package cz.adamec.timotej.snag.authentication.fe.driven.mobile.di
-
-import cz.adamec.timotej.snag.authentication.fe.driven.di.OIDC_REDIRECT_URI_QUALIFIER
-import cz.adamec.timotej.snag.configuration.fe.MobileRunConfig
-import org.koin.dsl.module
-
-val authenticationDrivenMobileModule =
-    module {
-        single(qualifier = OIDC_REDIRECT_URI_QUALIFIER) { MobileRunConfig.redirectUri }
-    }
