+++ feat/projects/fe/driving/impl/src/commonMain/kotlin/cz/adamec/timotej/snag/projects/fe/driving/impl/internal/projects/vm/ProjectsUiState.kt
@@ -18,7 +18,7 @@ import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
 
 @Immutable
-internal data class ProjectsUiState(
+data class ProjectsUiState(
     val projects: ImmutableList<AppProject> = persistentListOf(),
     val canCreateProject: Boolean = false,
 )
