+++ feat/projects/fe/driving/impl/src/commonMain/kotlin/cz/adamec/timotej/snag/projects/fe/driving/impl/internal/projects/vm/ProjectsViewModel.kt
@@ -30,7 +30,7 @@ import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.flow.update
 
-internal class ProjectsViewModel(
+class ProjectsViewModel(
     private val getProjectsUseCase: GetProjectsUseCase,
     private val canCreateProjectUseCase: CanCreateProjectUseCase,
 ) : ViewModel() {
