+++ feat/projects/fe/driving/impl/src/nonWebMain/kotlin/cz/adamec/timotej/snag/projects/fe/driving/impl/internal/projectDetails/vm/NonWebProjectDetailsViewModel.kt
@@ -40,7 +40,7 @@ import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import kotlin.uuid.Uuid
 
-internal class NonWebProjectDetailsViewModel(
+class NonWebProjectDetailsViewModel(
     projectId: Uuid,
     getProjectUseCase: GetProjectUseCase,
     deleteProjectUseCase: DeleteProjectUseCase,
