+++ feat/projects/fe/driving/impl/src/webMain/kotlin/cz/adamec/timotej/snag/projects/fe/driving/impl/internal/projectDetails/vm/WebProjectDetailsViewModel.kt
@@ -42,7 +42,7 @@ import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import kotlin.uuid.Uuid
 
-internal class WebProjectDetailsViewModel(
+class WebProjectDetailsViewModel(
     projectId: Uuid,
     getProjectUseCase: GetProjectUseCase,
     deleteProjectUseCase: DeleteProjectUseCase,
