+++ feat/projects/fe/driving/nonWear/build.gradle.kts
@@ -1,0 +1,36 @@
+/*
+ * Copyright (c) 2026 Timotej Adamec
+ * SPDX-License-Identifier: MIT
+ *
+ * This file is part of the thesis:
+ * "Multiplatform snagging system with code sharing maximisation"
+ *
+ * Czech Technical University in Prague
+ * Faculty of Information Technology
+ * Department of Software Engineering
+ */
+
+plugins {
+    alias(libs.plugins.snagDrivingFrontendMultiplatformModule)
+}
+
+kotlin {
+    sourceSets {
+        commonMain {
+            dependencies {
+                implementation(project(":feat:projects:fe:driving:impl"))
+                implementation(project(":feat:projects:fe:app:api"))
+                implementation(project(":feat:projects:business:model"))
+                implementation(project(":feat:structures:fe:app:api"))
+                implementation(project(":feat:structures:fe:driving:api"))
+                implementation(project(":feat:clients:fe:app:api"))
+                implementation(project(":feat:clients:fe:driving:api"))
+                implementation(project(":feat:inspections:fe:app:api"))
+                implementation(project(":feat:inspections:fe:driving:api"))
+                implementation(project(":feat:reports:fe:app:api"))
+                implementation(project(":feat:users:fe:app:api"))
+                implementation(project(":feat:users:fe:driving:api"))
+            }
+        }
+    }
+}
