+++ feat/projects/fe/driving/wear/build.gradle.kts
@@ -1,0 +1,26 @@
+/*
+ * Copyright (c) 2026 Timotej Adamec
+ * SPDX-License-Identifier: MIT
+ *
+ * This file is part of the thesis:
+ * "Multiplatform snagging system with code sharing maximisation"
+ *
+ * Czech Technical University in Prague
+ * Faculty of Information Technology
+ * Department of Software Engineering
+ */
+
+plugins {
+    alias(libs.plugins.snagDrivingFrontendMultiplatformModule)
+}
+
+kotlin {
+    sourceSets {
+        commonMain {
+            dependencies {
+                implementation(project(":feat:projects:fe:driving:impl"))
+                implementation(project(":feat:projects:fe:app:api"))
+            }
+        }
+    }
+}
