+++ settings.gradle.kts
@@ -71,6 +71,8 @@ include(":feat:projects:be:ports")
 include(":feat:projects:be:app:model")
 include(":feat:projects:fe:driving:api")
 include(":feat:projects:fe:driving:impl")
+include(":feat:projects:fe:driving:nonWear")
+include(":feat:projects:fe:driving:wear")
 include(":feat:shared:database:fe:api")
 include(":feat:shared:database:fe:impl")
 include(":feat:shared:database:fe:test")
@@ -212,7 +214,7 @@ include(":feat:authentication:fe:app:api")
 include(":feat:authentication:fe:app:impl")
 include(":feat:authentication:fe:ports")
 include(":feat:authentication:fe:driven:impl")
-include(":feat:authentication:fe:driven:mobile")
+include(":feat:authentication:fe:driven:nonWear")
 include(":feat:authentication:fe:driven:wear")
 include(":feat:authentication:fe:driven:test")
 include(":feat:authentication:fe:driving:api")
